test: consolidate duplicated cell-5/cell-7 coverage per review
Cell 7: relocate #1349's wait-landmark cases from selector-read.test.ts to
selector-wait.test.ts (the 1:1 topology location for selector-wait.ts),
replacing the weaker duplicate cell-7 cases added in the prior commit. The
relocated tests keep the stronger assertions (real computeTargetEvidence-
derived evidence, an initial no-match poll, observed-ancestry checks, and
the plain-timeout-vs-landmark-mismatch distinction).

Cell 5: the first case overlapped an existing later-alternative regression
in session-replay-target-classification.test.ts. Sharpened it (rather than
dropping it, since it is the only counterfactual-sensitive case for the
allowDisambiguation=false skip path) to isolate the branch the existing
regression's exact-tie fixture cannot reach, and paired it explicitly with
the second case as a same-fixture, flag-flipped contrast.
