+++ src/commands/interaction/runtime/selector-wait.test.ts
@@ -8,12 +8,14 @@ import {
   localCommandPolicy,
 } from '../../../runtime.ts';
 import { makeSnapshotState } from '../../../__tests__/test-utils/index.ts';
-import type { TargetAnnotationV1 } from '@agent-device/contracts/replay';
 import {
   createFakeClock,
   createSelectorDevice,
   selectorReadSnapshot,
 } from './__tests__/test-utils/index.ts';
+import { computeTargetEvidence } from '../../../daemon/session-target-evidence.ts';
+import { WAIT_LANDMARK_MISMATCH_REASON } from '../../../replay/target-identity-node.ts';
+import { AppError } from '@agent-device/kernel/errors';
 
 test('runtime focused selector waits against a full snapshot', async () => {
   const snapshot = makeSnapshotState([
@@ -65,117 +67,149 @@ test('runtime wait can use backend text search', async () => {
 });
 
 // ---------------------------------------------------------------------------
-// #1478 P5 step 2, cell 7: wait-landmark identity mismatch (#1349's
-// `recordedLandmark`). This is the root seam the future `resolveRecordedTarget`
-// port must preserve — a selector match alone is not enough; the wait only
-// reports success once SOME match carries the recorded landmark identity, and
-// a deadline with only impostor matches must surface the
-// `WAIT_LANDMARK_MISMATCH_REASON` refusal rather than either succeeding or an
-// undifferentiated timeout.
+// #1349 (relocated from `selector-read.test.ts` — this is the 1:1 topology
+// location for `selector-wait.ts`, and #1478 P5 step 2 cell 7's pin):
+// wait's in-loop landmark identity verification, threaded as
+// `target.recordedLandmark`. Polling semantics are preserved — a
+// same-selector impostor never aborts the wait; only the deadline turns
+// rejected candidates into the fail-closed landmark refusal
+// (`WAIT_LANDMARK_MISMATCH_REASON`), and a plain "the selector never matched
+// at all" timeout stays undifferentiated. This is the root seam the future
+// `resolveRecordedTarget` port operation must preserve.
 // ---------------------------------------------------------------------------
 
-function screenNodes(parentLabel: string) {
-  return [
-    {
-      index: 0,
-      depth: 0,
-      type: 'FrameLayout',
-      label: parentLabel,
-      rect: { x: 0, y: 0, width: 390, height: 844 },
-    },
+function landmarkScreen(parentLabel: string) {
+  return makeSnapshotState([
+    { index: 0, depth: 0, type: 'Other', label: parentLabel },
     {
       index: 1,
       depth: 1,
       parentIndex: 0,
-      type: 'TextView',
+      type: 'StaticText',
       label: 'Screen X',
-      rect: { x: 0, y: 100, width: 390, height: 40 },
+      rect: { x: 0, y: 0, width: 100, height: 20 },
     },
-  ];
+  ]);
 }
 
-function recordedLandmark(): TargetAnnotationV1 {
-  return {
-    role: 'textview',
-    label: 'Screen X',
-    ancestry: [{ role: 'framelayout', label: 'Detail Screen' }],
-    sibling: 0,
-    viewportOrder: 0,
-    verification: 'verified',
-  };
+function recordedLandmarkFor(snapshot: ReturnType<typeof landmarkScreen>) {
+  const node = snapshot.nodes[1]!;
+  const evidence = computeTargetEvidence(
+    { node, preActionNodes: snapshot.nodes },
+    { mode: 'landmark' },
+  );
+  assert.ok(evidence);
+  assert.equal(evidence.verification, 'verified');
+  return evidence;
 }
 
-test('P5 port cell 7: a landmark wait succeeds once a poll carries the recorded identity, not on the first same-selector match', async () => {
-  const clock = createFakeClock();
-  let calls = 0;
+function landmarkWaitDevice(captures: Array<ReturnType<typeof landmarkScreen>>) {
+  let call = 0;
+  const initial = captures[0]!;
   const device = createAgentDevice({
     backend: {
-      platform: 'android',
+      platform: 'ios',
       captureSnapshot: async () => {
-        calls += 1;
-        // First poll: an impostor screen — same selector match ("Screen X"),
-        // wrong ancestor label, so the recorded landmark's ancestry prefix
-        // does not match. Second poll: the real destination screen.
-        const nodes = makeSnapshotState(
-          calls === 1 ? screenNodes('List Screen') : screenNodes('Detail Screen'),
-        );
-        return { snapshot: nodes };
+        const snapshot = captures[Math.min(call, captures.length - 1)]!;
+        call += 1;
+        return { snapshot };
       },
     } satisfies AgentDeviceBackend,
     artifacts: createLocalArtifactAdapter(),
-    sessions: createMemorySessionStore([{ name: 'default', snapshot: makeSnapshotState([]) }]),
+    sessions: createMemorySessionStore([{ name: 'default', snapshot: initial }]),
     policy: localCommandPolicy(),
-    clock,
+    clock: createFakeClock(),
   });
+  return device;
+}
+
+test('runtime wait keeps polling past a same-selector impostor and succeeds on the recorded landmark', async () => {
+  const recordTime = landmarkScreen('Detail Screen');
+  const recorded = recordedLandmarkFor(recordTime);
+  const impostor = landmarkScreen('List Screen');
+  const empty = makeSnapshotState([{ index: 0, depth: 0, type: 'Other', label: 'Loading' }]);
+  const device = landmarkWaitDevice([empty, impostor, landmarkScreen('Detail Screen')]);
 
   const result = await device.selectors.wait({
     session: 'default',
     target: {
       kind: 'selector',
       selector: 'label="Screen X"',
-      timeoutMs: 5000,
-      recordedLandmark: recordedLandmark(),
+      timeoutMs: 10_000,
+      recordedLandmark: recorded,
     },
   });
 
   assert.equal(result.kind, 'selector');
-  if (result.kind !== 'selector') return;
+  if (result.kind !== 'selector') throw new Error('unreachable');
+  // Two rejected polls (absent, then impostor) before the landmark appeared.
+  assert.equal(result.waitedMs >= 600, true);
   assert.equal(result.node?.label, 'Screen X');
-  assert.ok(calls >= 2, `must not report success on the impostor's own poll, got ${calls} polls`);
-  assert.ok(result.waitedMs > 0, 'must have advanced past the impostor poll');
+  assert.equal(result.preActionNodes?.length, 2);
 });
 
-test('P5 port cell 7: a landmark wait refuses at the deadline with WAIT_LANDMARK_MISMATCH_REASON when every poll is an impostor', async () => {
-  const clock = createFakeClock();
-  const device = createAgentDevice({
-    backend: {
-      platform: 'android',
-      captureSnapshot: async () => ({ snapshot: makeSnapshotState(screenNodes('List Screen')) }),
-    } satisfies AgentDeviceBackend,
-    artifacts: createLocalArtifactAdapter(),
-    sessions: createMemorySessionStore([{ name: 'default', snapshot: makeSnapshotState([]) }]),
-    policy: localCommandPolicy(),
-    clock,
-  });
+test('runtime wait fails closed at the deadline when only impostors matched the selector', async () => {
+  const recorded = recordedLandmarkFor(landmarkScreen('Detail Screen'));
+  const device = landmarkWaitDevice([landmarkScreen('List Screen')]);
+
+  const error = await device.selectors
+    .wait({
+      session: 'default',
+      target: {
+        kind: 'selector',
+        selector: 'label="Screen X"',
+        timeoutMs: 1000,
+        recordedLandmark: recorded,
+      },
+    })
+    .then(
+      () => undefined,
+      (thrown: unknown) => thrown,
+    );
+
+  assert.ok(error instanceof AppError);
+  assert.equal(error.details?.reason, WAIT_LANDMARK_MISMATCH_REASON);
+  assert.equal(error.details?.matchCount, 1);
+  const observed = error.details?.observed as { role: string; label?: string };
+  assert.equal(observed.label, 'Screen X');
+  const ancestry = error.details?.observedAncestry as Array<{ role: string; label?: string }>;
+  assert.equal(ancestry[0]?.label, 'List Screen');
+});
+
+test('runtime wait with a recorded landmark keeps the plain timeout when the selector never matched', async () => {
+  const recorded = recordedLandmarkFor(landmarkScreen('Detail Screen'));
+  const empty = makeSnapshotState([{ index: 0, depth: 0, type: 'Other', label: 'Loading' }]);
+  const device = landmarkWaitDevice([empty]);
 
   await assert.rejects(
-    () =>
-      device.selectors.wait({
-        session: 'default',
-        target: {
-          kind: 'selector',
-          selector: 'label="Screen X"',
-          timeoutMs: 250,
-          recordedLandmark: recordedLandmark(),
-        },
-      }),
-    (error: unknown) => {
-      const details = (error as { details?: { reason?: string; matchCount?: number } }).details;
-      assert.equal(details?.reason, 'wait_landmark_identity_mismatch');
-      // The selector itself matched (one impostor node) — this must not be
-      // reported as an ordinary selector-not-found timeout.
-      assert.equal(details?.matchCount, 1);
+    device.selectors.wait({
+      session: 'default',
+      target: {
+        kind: 'selector',
+        selector: 'label="Screen X"',
+        timeoutMs: 1000,
+        recordedLandmark: recorded,
+      },
+    }),
+    (thrown: unknown) => {
+      assert.ok(thrown instanceof AppError);
+      assert.match(thrown.message, /wait timed out for selector/);
+      assert.equal(thrown.details?.reason, undefined);
       return true;
     },
   );
+});
+
+test('runtime wait without a recorded landmark returns the satisfying match for record-time evidence', async () => {
+  const device = landmarkWaitDevice([landmarkScreen('Detail Screen')]);
+
+  const result = await device.selectors.wait({
+    session: 'default',
+    target: { kind: 'selector', selector: 'label="Screen X"', timeoutMs: 1000 },
+  });
+
+  assert.equal(result.kind, 'selector');
+  if (result.kind !== 'selector') throw new Error('unreachable');
+  assert.equal(result.node?.label, 'Screen X');
+  assert.equal(result.preActionNodes?.length, 2);
 });
